pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "pillow>=11.1.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.6",
]
